SmartLibrarian/requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
typer>=0.12.0
rich>=13.7.1
